+++ fiddles/vue/fiddle-0002-AgGridSparklineTs/src/App.vue
@@ -1,0 +1,70 @@
+<script setup lang="ts">
+import {  RouterView } from 'vue-router'
+import TopNav from './components/TopNav.vue'
+</script>
+
+
+<template>
+ 
+ <TopNav title="Vue Fiddle #2 ~ AgGrid Sparkline (TS)" href="https://github.com/bradyhouse/house/tree/master/fiddles/vue/fiddle-0002-AgGridSparklineTs"></TopNav>
+  <RouterView :key="$route.fullPath" />
+</template>
+
+<style scoped>
+header {
+  line-height: 1.5;
+  max-height: 100vh;
+}
+
+.logo {
+  display: block;
+  margin: 0 auto 2rem;
+}
+
+nav {
+  width: 100%;
+  font-size: 12px;
+  text-align: center;
+}
+
+nav a.router-link-exact-active {
+  color: var(--color-text);
+}
+
+nav a.router-link-exact-active:hover {
+  background-color: transparent;
+}
+
+nav a {
+  display: inline-block;
+  padding: 0 1rem;
+  border-left: 1px solid var(--color-border);
+}
+
+nav a:first-of-type {
+  border: 0;
+}
+
+@media (min-width: 1024px) {
+  header {
+    display: flex;
+    place-items: center;
+    padding-right: calc(var(--section-gap) / 2);
+  }
+
+  .logo {
+    margin: 0 2rem 0 0;
+  }
+
+  header .wrapper {
+    display: flex;
+    place-items: flex-start;
+    flex-wrap: wrap;
+  }
+
+  nav {
+    text-align: left;
+    font-size: 1rem;
+  }
+}
+</style>