+++ fiddles/vue/fiddle-0002-AgGridSparklineTs/src/router/index.ts
@@ -1,0 +1,28 @@
+import { createRouter, createWebHistory } from 'vue-router'
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes: [
+    {
+      path: '/',
+      name: 'fiddle',
+      component: () => import('../views/FiddleView.vue')
+    },
+
+    {
+      path: '/docs',
+      name: 'docs',
+      component: () => import('../views/HomeView.vue')
+    },
+
+    {
+      path: '/about',
+      name: 'about',
+      // route level code-splitting
+      // this generates a separate chunk (About.[hash].js) for this route
+      // which is lazy-loaded when the route is visited.
+      component: () => import('../views/AboutView.vue')
+    },
+    ]
+})
+
+export default router